+++ src/irmin-pack/ext.mli
@@ -17,5 +17,5 @@
 module Maker
     (_ : Version.S)
     (_ : Conf.S)
-    (N : Irmin.Private.Node.Maker)
-    (CT : Irmin.Private.Commit.Maker) : S.Maker with type info = CT.Info.t
+    (N : Irmin.Node.Maker)
+    (CT : Irmin.Commit.Maker) : S.Maker with type info = CT.Info.t