+++ src/irmin-pack/mem/irmin_pack_mem.mli
@@ -18,5 +18,4 @@
     backend, intended for users that must be interoperable with the
     idiosyncrasies of the persistent implementation. *)
 
-module Maker (_ : Irmin.Private.Node.Maker) (_ : Irmin.Private.Commit.Maker) :
-  Irmin_pack.Maker
+module Maker (_ : Irmin.Node.Maker) (_ : Irmin.Commit.Maker) : Irmin_pack.Maker